+++ packages/browser-wallet/src/popup/pages/Settings/Settings.tsx
@@ -9,6 +9,10 @@ import { Theme } from '@shared/storage/types';
 import SunIcon from '@assets/svg/sun.svg';
 import MoonIcon from '@assets/svg/moon.svg';
 import { ToggleCheckbox } from '@popup/shared/Form/ToggleCheckbox';
+import { popupMessageHandler } from '@popup/shared/message-handler';
+import { InternalMessageType } from '@concordium/browser-wallet-message-hub';
+import Button from '@popup/shared/Button';
+import { isFullscreenWindow } from '@popup/shared/window-helpers';
 
 function LightDarkModeToggle() {
     const { t } = useTranslation('settings');
@@ -47,6 +51,17 @@ export default function Settings() {
                 <Link className="settings-page__link" to={absoluteRoutes.home.settings.seedPhrase.path}>
                     {t('seedPhrase')}
                 </Link>
+                {!isFullscreenWindow && (
+                    <Button
+                        clear
+                        className="settings-page__link settings-page__link--button"
+                        onClick={() => {
+                            popupMessageHandler.sendInternalMessage(InternalMessageType.OpenFullscreen);
+                        }}
+                    >
+                        {t('fullscreenWallet')}
+                    </Button>
+                )}
                 <Link className="settings-page__link" to={absoluteRoutes.home.settings.about.path}>
                     {t('about')}
                 </Link>